requirements.txt
scapy>=2.5.0
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.10.0
netifaces>=0.11.0
psutil>=5.9.0
python-dotenv>=1.0.0
